+++ flotaydepositos/src/main/java/utnfrc/isi/backend/FlotaYDepositos/dto/CreateDepositoRequestDTO.java
@@ -1,5 +1,7 @@
 package utnfrc.isi.backend.FlotaYDepositos.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -10,10 +12,10 @@ public record CreateDepositoRequestDTO(
     @NotBlank(message = "La dirección no puede estar vacía")
     String direccion,
     @NotNull(message = "La latitud no puede ser nula")
-    Double latitud,
+    BigDecimal latitud,
     @NotNull(message = "La longitud no puede ser nula")
-    Double longitud,
+    BigDecimal longitud,
     @NotNull(message = "El costo de estadía no puede ser nulo")
     @Positive(message = "El costo de estadía debe ser un valor positivo")
-    Double costoEstadiaDiario
+    BigDecimal costoEstadiaDiario
 ) {}
